Lifted state up. Added refresh button
