+++ src/visualizer/index.tsx
@@ -5,10 +5,53 @@ import NodeMap from './nodeMap'
 import { PacketSortOptions as PacketSort, PACKET_PAGE_SIZE} from '../common/Constants'
 import PacketViewSettingsModal from './PacketViewSettingsModal'
 import PacketViewSettingsState from './PacketViewSettingsState'
+import APIUtil from '../utilities/APIutils'
+import PacketState from './packetContainer/PacketState'
 import './index.css'
 
 function Visualizer() {
     const params = useParams()
+
+    const api = new APIUtil()
+
+    // Packet retrieval and infinite list
+    let [page, setPage] = useState(0)
+    let [packetList, setPacketList]: Array<any> = useState([])
+    let [hasMorePackets, setHasMorePackets] = useState(true)
+    const parsePackets = (packets: PacketState[]) => packets.map((packet) => {
+        return (
+            <tr>
+                <td>{packet.timestamp}</td>
+                <td>{packet.id}</td>
+                <td>{packet.type}</td>
+                <td>{packet.data}</td>
+            </tr>
+        )
+    })
+    const fetchPackets = () => {
+        const newData = api.getPackets(page)
+        const newPackets = parsePackets(newData)
+
+        if (newData.length > 0) {
+            setPacketList(packetList.concat(newPackets))
+            setPage(page + 1)
+        } else {
+            setHasMorePackets(false)
+        }
+    }
+    const refreshPackets = () => {
+        setPage(0)
+        const newData = api.getPackets(page)
+        const newPackets = parsePackets(newData)
+        setPacketList(newPackets)
+
+        if (newData.length > 0) {
+            setPage(page + 1)
+        }
+        else {
+            setHasMorePackets(false)
+        }
+    }
 
     // Modal for changing packet view settings
     let [isShownPacketsModal, setIsShownPacketsModal] = useState(true)
@@ -33,7 +76,12 @@ function Visualizer() {
             <h1 className='visualizer-title'>{params.projectId}</h1>
             <div className='visualizer-content'>
                 <div className='packet-container-content'>
-                    <PacketContainer></PacketContainer>
+                    <PacketContainer
+                    fetchData={fetchPackets}
+                    hasMore={hasMorePackets}
+                    packetList={packetList}
+                    refresh={refreshPackets}
+                    />
                 </div>
                 <div className='node-map-container-content'>
                     <NodeMap></NodeMap>
